+++ changelog.md
@@ -40,6 +40,7 @@
  ### Lua API
    * unit:transform() now takes an optional variation parameter
    * Support side.variables for access to side variables, similar to unit.variables
+   * New read-write keys in unit userdata: ellipse, halo, description, renamable
  ### WML engine
    * Support upkeep in StandardUnitFilter
    * [effect]apply_to=variation now supports heal_full